requirements.txt
google-adk
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
